pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "localscript"
version = "0.1.0"
description = "Local agentic system for Lua code generation and validation"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "openai>=1.55",
    "faiss-cpu>=1.9",
    "sentence-transformers>=3.3",
    "structlog>=24.4",
    "aiosqlite>=0.20",
    "pyyaml>=6.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
